backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pymongo==4.10.1
pypdf==5.1.0
python-docx==1.1.2
python-multipart==0.0.19
fpdf2==2.8.1
